+++ packages/in_app_purchase/in_app_purchase_storekit/example/shared/RunnerTests/InAppPurchaseStoreKit2PluginTests.swift
@@ -60,6 +60,35 @@ final class InAppPurchase2PluginTests: XCTestCase {
     XCTAssertEqual(testProductMsg, fetchedProductMsg)
   }
 
+  func testGetTransactions() async throws {
+    let purchaseExpectation = self.expectation(description: "Purchase should succeed")
+    let transactionExpectation = self.expectation(
+      description: "Getting transactions should succeed")
+
+    plugin.purchase(id: "consumable", options: nil) { result in
+      switch result {
+      case .success:
+        purchaseExpectation.fulfill()
+      case .failure(let error):
+        XCTFail("Purchase should NOT fail. Failed with \(error)")
+      }
+    }
+
+    await fulfillment(of: [purchaseExpectation], timeout: 5)
+
+    plugin.transactions {
+      result in
+      switch result {
+      case .success(let transactions):
+        XCTAssert(transactions.count == 1)
+        transactionExpectation.fulfill()
+      case .failure(let error):
+        XCTFail("Getting transactions should NOT fail. Failed with \(error)")
+      }
+    }
+    await fulfillment(of: [transactionExpectation], timeout: 5)
+  }
+
   func testGetDiscountedProducts() async throws {
     let expectation = self.expectation(description: "products successfully fetched")
 
@@ -91,7 +120,7 @@ final class InAppPurchase2PluginTests: XCTestCase {
       case .success(let productMessages):
         fetchedProductMsg = productMessages
         expectation.fulfill()
-      case .failure(_):
+      case .failure:
         XCTFail("Products should be successfully fetched")
       }
     }
@@ -110,7 +139,7 @@ final class InAppPurchase2PluginTests: XCTestCase {
 
     plugin.products(identifiers: ["subscription_silver"]) { result in
       switch result {
-      case .success(_):
+      case .success:
         XCTFail("This `products` call should not succeed")
       case .failure(let error):
         expectation.fulfill()
@@ -127,7 +156,7 @@ final class InAppPurchase2PluginTests: XCTestCase {
     let expectation = self.expectation(description: "Purchase request should succeed")
     plugin.purchase(id: "consumable", options: nil) { result in
       switch result {
-      case .success(let purchaseResult):
+      case .success:
         expectation.fulfill()
       case .failure(let error):
         XCTFail("Purchase should NOT fail. Failed with \(error)")
@@ -143,7 +172,7 @@ final class InAppPurchase2PluginTests: XCTestCase {
     let expectation = self.expectation(description: "products request should fail")
     plugin.purchase(id: "consumable", options: nil) { result in
       switch result {
-      case .success(_):
+      case .success:
         XCTFail("Purchase should NOT suceed.")
       case .failure(let error):
         XCTAssertEqual(
@@ -162,7 +191,7 @@ final class InAppPurchase2PluginTests: XCTestCase {
     let expectation = self.expectation(description: "Purchase request should succeed")
     plugin.purchase(id: "consumable", options: nil) { result in
       switch result {
-      case .success(_):
+      case .success:
         XCTFail("Purchase should NOT suceed.")
       case .failure(let error):
         XCTAssertEqual(error.localizedDescription, "Item Unavailable")
@@ -176,7 +205,7 @@ final class InAppPurchase2PluginTests: XCTestCase {
     let expectation = self.expectation(description: "products request should fail")
     plugin.purchase(id: "invalid_product", options: nil) { result in
       switch result {
-      case .success(_):
+      case .success:
         XCTFail("Purchase should NOT suceed.")
       case .failure(let error):
         let pigeonError = error as! PigeonError
@@ -192,7 +221,7 @@ final class InAppPurchase2PluginTests: XCTestCase {
     let expectation = self.expectation(description: "Purchase request should succeed")
     plugin.purchase(id: "subscription_discounted", options: nil) { result in
       switch result {
-      case .success(let purchaseResult):
+      case .success:
         expectation.fulfill()
       case .failure(let error):
         XCTFail("Purchase should NOT fail. Failed with \(error)")
@@ -205,7 +234,7 @@ final class InAppPurchase2PluginTests: XCTestCase {
     let expectation = self.expectation(description: "Purchase request should succeed")
     plugin.purchase(id: "subscription_discounted", options: nil) { result in
       switch result {
-      case .success(let purchaseResult):
+      case .success:
         expectation.fulfill()
       case .failure(let error):
         XCTFail("Purchase should NOT fail. Failed with \(error)")
@@ -218,7 +247,7 @@ final class InAppPurchase2PluginTests: XCTestCase {
     let expectation = self.expectation(description: "Purchase request should succeed")
     plugin.purchase(id: "consumable_discounted", options: nil) { result in
       switch result {
-      case .success(let purchaseResult):
+      case .success:
         expectation.fulfill()
       case .failure(let error):
         XCTFail("Purchase should NOT fail. Failed with \(error)")
@@ -233,7 +262,7 @@ final class InAppPurchase2PluginTests: XCTestCase {
 
     plugin.purchase(id: "subscription_silver", options: nil) { result in
       switch result {
-      case .success(_):
+      case .success:
         purchaseExpectation.fulfill()
       case .failure(let error):
         XCTFail("Purchase should NOT fail. Failed with \(error)")
@@ -249,5 +278,33 @@ final class InAppPurchase2PluginTests: XCTestCase {
     }
 
     await fulfillment(of: [restoreExpectation, purchaseExpectation], timeout: 5)
+  }
+
+  func testFinishTransaction() async throws {
+    let purchaseExpectation = self.expectation(description: "Purchase should succeed")
+    let finishExpectation = self.expectation(description: "Finishing purchase should succeed")
+
+    plugin.purchase(id: "consumable", options: nil) { result in
+      switch result {
+      case .success(let purchase):
+        purchaseExpectation.fulfill()
+      case .failure(let error):
+        XCTFail("Purchase should NOT fail. Failed with \(error)")
+      }
+    }
+
+    await fulfillment(of: [purchaseExpectation], timeout: 5)
+
+    // id should always be 0 as it is the first purchase
+    plugin.finish(id: 0) { result in
+      switch result {
+      case .success():
+        finishExpectation.fulfill()
+      case .failure(let error):
+        XCTFail("Finish purchases should NOT fail. Failed with \(error)")
+      }
+    }
+
+    await fulfillment(of: [finishExpectation], timeout: 5)
   }
 }